Update Buttons Flexbox
After looking at #15 - I decided to play with the flexbox setup on the buttons a bit.  This small change just keeps the buttons naturally sized in-lieu of growing to fit the entire space on the widget screen, and adds `flex-wrap` to handle any situations where long translations would cause the buttons to get too large for one line.